feat: implement custom 404 page and update header links for improved navigation
