+++ components/Header.vue
@@ -53,8 +53,10 @@ const handleLogout = () => {
         aria-labelledby="hs-navbar-example-collapse">
         <div class="block md:flex md:justify-center md:items-center md:gap-4">
           <div class="flex flex-col gap-5 mt-5 sm:flex-row sm:items-center sm:justify-end sm:mt-0 sm:ps-5">
-            <NuxtLink :to="{ name: 'index', query: { 'face-swap-app': true } }"
-              class="text-white font-medium focus:outline-hidden">
+            <NuxtLink to="/" class="text-white font-medium focus:outline-hidden">
+              Home
+            </NuxtLink>
+            <NuxtLink to="/face-swap" class="text-white font-medium focus:outline-hidden">
               Face Swap Photo
             </NuxtLink>
           </div>
