+++ nuxt.config.js
@@ -82,7 +82,8 @@ export default defineNuxtConfig({
         },
       ],
       bodyAttrs: {
-        class: "relative flex flex-col px-4 md:px-0 pt-[5rem] bg-slate-900 text-neutral-50",
+        class:
+          "relative flex flex-col px-4 md:px-0 pt-[5rem] bg-linear-to-b from-violet-600/10 via-transparent text-neutral-50",
       },
     },
   },
