+++ pages/terms-of-service.vue
@@ -1,7 +1,7 @@
 <template>
   <div class="max-w-4xl mx-auto p-6">
     <h1 class="text-3xl font-bold text-center mb-6">Terms of Service</h1>
-    <p class="text-sm text-gray-500 text-center mb-6">Last Updated: Apr 03, 2025</p>
+    <p class="text-sm text-gray-500 mb-6">Last Updated: Apr 03, 2025</p>
 
     <section class="mb-6">
       <p>
